Fix verbose log for subcommands
