+++ codechecker/CodeChecker.py
@@ -20,6 +20,7 @@ import shared
 from cmdline_client import cmd_line_client
 from codechecker_lib import arg_handler
 from codechecker_lib import util
+from codechecker_lib import logger
 from codechecker_lib.logger import LoggerFactory
 from codechecker_lib.analyzers import analyzer_types
 
@@ -147,17 +148,6 @@ def add_analyzer_arguments(parser):
                         "without modification.")
 
 
-def add_verbose_arguments(parser):
-    """
-    Verbosity level arguments.
-    """
-    parser.add_argument('--verbose', type=str, dest='verbose',
-                        choices=['info', 'debug', 'debug_analyzer'],
-                        default='info',
-                        help='Set verbosity level.')
-
-
-# ------------------------------------------------------------------------------
 def main():
     """
     CodeChecker main command line.
@@ -301,7 +291,7 @@ Build command which is used to build the project.''')
 
         add_analyzer_arguments(check_parser)
         add_database_arguments(check_parser)
-        add_verbose_arguments(check_parser)
+        logger.add_verbose_arguments(check_parser)
         check_parser.set_defaults(func=arg_handler.handle_check)
 
         # --------------------------------------
@@ -351,7 +341,7 @@ Build command which is used to build the project.''')
                                         'This is useful when you'
                                         'do cross-compilation.')
         add_analyzer_arguments(qcheck_parser)
-        add_verbose_arguments(qcheck_parser)
+        logger.add_verbose_arguments(qcheck_parser)
         qcheck_parser.set_defaults(func=arg_handler.handle_quickcheck)
 
         # --------------------------------------
@@ -377,7 +367,7 @@ Build command which is used to build the project.''')
                                default=argparse.SUPPRESS,
                                required=True, help='Build command.')
 
-        add_verbose_arguments(logging_p)
+        logger.add_verbose_arguments(logging_p)
         logging_p.set_defaults(func=arg_handler.handle_log)
 
         # --------------------------------------
@@ -397,7 +387,7 @@ Build command which is used to build the project.''')
                                'should be listed.\nCurrently supported '
                                'analyzers:\n' + analyzers)
 
-        add_verbose_arguments(checker_p)
+        logger.add_verbose_arguments(checker_p)
         checker_p.set_defaults(func=arg_handler.handle_list_checkers)
 
         # --------------------------------------
@@ -436,14 +426,13 @@ Build command which is used to build the project.''')
                                    required=False, help='Server address.')
 
         add_database_arguments(server_parser)
-        add_verbose_arguments(server_parser)
+        logger.add_verbose_arguments(server_parser)
         server_parser.set_defaults(func=arg_handler.handle_server)
 
         # --------------------------------------
         # Cmd_line.
         cmd_line_parser = subparsers.add_parser('cmd',
                                                 help='Command line client')
-        add_verbose_arguments(cmd_line_parser)
         cmd_line_client.register_client_command_line(cmd_line_parser)
 
         # --------------------------------------
@@ -468,7 +457,7 @@ Build command which is used to build the project.''')
                                   help='Overwrite already generated files.')
 
         add_database_arguments(debug_parser)
-        add_verbose_arguments(debug_parser)
+        logger.add_verbose_arguments(debug_parser)
         debug_parser.set_defaults(func=arg_handler.handle_debug)
 
         # --------------------------------------
@@ -516,7 +505,7 @@ Build command which is used to build the project.''')
                                        'name already exists.')
 
         add_database_arguments(plist_parser)
-        add_verbose_arguments(plist_parser)
+        logger.add_verbose_arguments(plist_parser)
         plist_parser.set_defaults(func=arg_handler.handle_plist)
 
         # --------------------------------------
@@ -525,7 +514,7 @@ Build command which is used to build the project.''')
                                                help='Print package version '
                                                     'information.')
         version_parser.set_defaults(func=arg_handler.handle_version_info)
-        add_verbose_arguments(version_parser)
+        logger.add_verbose_arguments(version_parser)
 
         args = parser.parse_args()
         LoggerFactory.set_log_level(args.verbose)
